Issue #SB-22862 feat: Collection CSV Hierarchy feature - code review changes implemented
