+++ course-mw/collection-actors/pom.xml
@@ -18,8 +18,8 @@
 		<scoverage.plugin.version>1.4.1</scoverage.plugin.version>
 	</properties>
 	<dependencies>
-		<dependency>
-			<groupId>ch.qos.logback</groupId>
+		<!-- <dependency>
+			 <groupId>ch.qos.logback</groupId>
 			<artifactId>logback-classic</artifactId>
 			<version>1.2.3</version>
 		</dependency>
@@ -27,7 +27,7 @@
 			<groupId>net.logstash.logback</groupId>
 			<artifactId>logstash-logback-encoder</artifactId>
 			<version>6.3</version>
-		</dependency>
+		</dependency> -->
 		<dependency>
 			<groupId>org.sunbird</groupId>
 			<artifactId>course-actors-common</artifactId>
